Moved auth library to shared area
