+++ libs/accruals/shell/src/lib/components/navbar/navbar.component.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterLink } from '@angular/router';
-import { AfAuthenticationService } from '@tfx-accruals/accruals/util/af-authentication';
+import { AfAuthenticationService } from '@tfx-accruals/shared/util/af-authentication';
 import { map } from 'rxjs';
 import { HamburgerButtonComponent } from './hamburger-button.component';
 import { LogoComponent } from './logo.component';
