+++ libs/shared/util/af-authentication/README.md
@@ -1,3 +1,3 @@
-# accruals-util-af-authentication
+# shared-util-af-authentication
 
 This library was generated with [Nx](https://nx.dev).
