+++ libs/shared/util/af-authentication/tsconfig.json
