+++ libs/shared/util/af-authentication/tsconfig.lib.json
